+++ lib/get-package-manager.d.ts
@@ -6,4 +6,4 @@ export type PackageManager = 'npm' | 'pnpm' | 'yarn';
  * Get the name of the preferred package manager from the `package.json` engines
  * property. Defaults to "npm".
  */
-export default function getPackageManager(engines?: PackageJson.engines): PackageManager;
+export default function getPackageManager(pkg?: PackageJson): PackageManager;