+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+[Unreleased]
+------------
+### Changed
+- Now takes `config.pkg.packageManager` into account in addition to `config.pkg.engines` when determining the package manager. See [here](https://nodejs.org/docs/v20.18.0/api/packages.html#packagemanager) for reference.
+
 [1.1.0] - 2024-10-11
 --------------------
 ### Added